@@ -3,12 +3,12 @@ plugins {
 }
 
 android {
-    compileSdkVersion 29
-
+    compileSdkVersion 34
+    namespace 'com.palm360.palmgrading'
     defaultConfig {
         applicationId "com.palm360.palmgrading"
         minSdkVersion 21
-        targetSdkVersion 29
+        targetSdkVersion 34
         versionCode 1
         versionName "1.0"
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
@@ -51,7 +51,9 @@ android {
         sourceCompatibility JavaVersion.VERSION_1_8
         targetCompatibility JavaVersion.VERSION_1_8
     }
-
+    buildFeatures {
+        buildConfig = true
+    }
     packagingOptions {
         exclude 'META-INF/ASL2.0'
         exclude 'META-INF/DEPENDENCIES'
