@@ -31,13 +31,23 @@ public class Palm3FoilDatabase extends SQLiteOpenHelper {
     private static SQLiteDatabase mSqLiteDatabase = null;
     private Context mContext;
     File rootDirectory;
+    private static String FULL_DB_PATH;
+
 
     public Palm3FoilDatabase(Context context) {
         super(context, DATABASE_NAME, null, DATA_VERSION);
         this.mContext = context;
+
         File dbDirectory = new File(CommonUtils.get3FFileRootPath() + "3F_Database");
+        if (!dbDirectory.exists()) {
+            boolean created = dbDirectory.mkdirs();
+            Log.d(LOG_TAG, "DB directory created: " + created);
+        }
+
         DB_PATH = dbDirectory.getAbsolutePath() + File.separator;
-        Log.v("The Database Path", DB_PATH);
+        FULL_DB_PATH = DB_PATH + DATABASE_NAME;
+
+        Log.v("The Database Path", FULL_DB_PATH);
     }
 
     public static synchronized Palm3FoilDatabase getPalm3FoilDatabase(Context context) {
@@ -47,15 +57,12 @@ public class Palm3FoilDatabase extends SQLiteOpenHelper {
             }
             return palm3FoilDatabase;
         }
-
     }
 
     public static SQLiteDatabase openDataBaseNew() throws SQLException {
-        // Open the database
-        if (mSqLiteDatabase == null) {
-            mSqLiteDatabase = SQLiteDatabase.openDatabase(DB_PATH, null, SQLiteDatabase.OPEN_READWRITE | SQLiteDatabase.CREATE_IF_NECESSARY | SQLiteDatabase.NO_LOCALIZED_COLLATORS);
-        } else if (!mSqLiteDatabase.isOpen()) {
-            mSqLiteDatabase = SQLiteDatabase.openDatabase(DB_PATH, null, SQLiteDatabase.OPEN_READWRITE | SQLiteDatabase.CREATE_IF_NECESSARY | SQLiteDatabase.NO_LOCALIZED_COLLATORS);
+        if (mSqLiteDatabase == null || !mSqLiteDatabase.isOpen()) {
+            mSqLiteDatabase = SQLiteDatabase.openDatabase(FULL_DB_PATH, null,
+                    SQLiteDatabase.OPEN_READWRITE | SQLiteDatabase.CREATE_IF_NECESSARY | SQLiteDatabase.NO_LOCALIZED_COLLATORS);
         }
         return mSqLiteDatabase;
     }
@@ -64,7 +71,6 @@ public class Palm3FoilDatabase extends SQLiteOpenHelper {
         InputStream in = new FileInputStream(src);
         OutputStream out = new FileOutputStream(dst);
 
-        // Transfer bytes from in to out
         byte[] buf = new byte[1024];
         int len;
         while ((len = in.read(buf)) > 0) {
@@ -74,39 +80,44 @@ public class Palm3FoilDatabase extends SQLiteOpenHelper {
         out.close();
     }
 
-    /* create an empty database if data base is not existed */
     public void createDataBase() throws IOException {
-        boolean dbExist = checkDataBase();
-
-        if (dbExist) {
-            //do nothing - database already exist
-        } else {
+        if (!checkDataBase()) {
             try {
-                copyDataBase();
+                copyDataBase(); // You may need to define this or clarify if you're copying from assets
                 Log.v("dbcopied:::", "true");
-            } catch (SQLiteException ex) {
-                ex.printStackTrace();
-                throw new Error("Error copying database");
-            } catch (IOException e) {
+            } catch (Exception e) {
                 e.printStackTrace();
                 throw new Error("Error copying database");
             }
+
             try {
-                openDataBase();
-            } catch (SQLiteException ex) {
-                ex.printStackTrace();
-                throw new Error("Error opening database");
+                openDataBaseNew();
             } catch (Exception e) {
                 e.printStackTrace();
                 throw new Error("Error opening database");
             }
         }
-
     }
 
-    /* checking the data base is existed or not */
-    /* return true if existed else return false */
     private boolean checkDataBase() {
+        SQLiteDatabase checkDB = null;
+        try {
+            Log.e("DB Check Path", FULL_DB_PATH);
+            checkDB = SQLiteDatabase.openDatabase(FULL_DB_PATH, null, SQLiteDatabase.OPEN_READONLY);
+        } catch (SQLiteException e) {
+            Log.w(LOG_TAG, "Database does not exist yet.");
+        }
+
+        if (checkDB != null) {
+            checkDB.close();
+            return true;
+        }
+
+        return false;
+    }
+
+
+    /*  private boolean checkDataBase() {
         boolean dataBaseExisted = false;
         try {
             String check_Path = DB_PATH + DATABASE_NAME;
@@ -117,7 +128,7 @@ public class Palm3FoilDatabase extends SQLiteOpenHelper {
         }
         return mSqLiteDatabase != null;
     }
-
+*/
     private void copyFile(InputStream in, OutputStream out) throws IOException {
         byte[] buffer = new byte[1024];
         int read;
@@ -129,26 +140,13 @@ public class Palm3FoilDatabase extends SQLiteOpenHelper {
     private void copyDataBase() throws IOException {
         File dbDir = new File(DB_PATH);
         if (!dbDir.exists()) {
-            dbDir.mkdirs();
+            dbDir.mkdir();
 
         }
         InputStream myInput = mContext.getAssets().open(DATABASE_NAME);
         OutputStream myOutput = new FileOutputStream(DB_PATH + DATABASE_NAME);
         copyFile(myInput, myOutput);
 
-    }
-
-    /* Open the database */
-    public void openDataBase() throws SQLException {
-
-        String check_Path = DB_PATH + DATABASE_NAME;
-        if (mSqLiteDatabase != null) {
-            mSqLiteDatabase.close();
-            mSqLiteDatabase = null;
-            mSqLiteDatabase = SQLiteDatabase.openDatabase(check_Path, null, SQLiteDatabase.OPEN_READWRITE);
-        } else {
-            mSqLiteDatabase = SQLiteDatabase.openDatabase(check_Path, null, SQLiteDatabase.OPEN_READWRITE);
-        }
     }
 
     @Override
@@ -160,12 +158,6 @@ public class Palm3FoilDatabase extends SQLiteOpenHelper {
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
 
     }
-
-    public void openDatabase() {
-        String dbPath = mContext.getDatabasePath(DATABASE_NAME).getPath();
-        if(mSqLiteDatabase != null && mSqLiteDatabase.isOpen()) {
-            return;
-        }
-        mSqLiteDatabase = SQLiteDatabase.openDatabase(dbPath, null, SQLiteDatabase.OPEN_READWRITE);
-    }
 }
+
+
